gazebo_plugins was renamed to gazebo_plugins_auvsim
