+++ src/auv_gazebo/launch/pool.launch
@@ -2,7 +2,7 @@
 <launch>
 
   <arg name="world" default="empty"/> 
-  <arg name="paused" default="true"/>
+  <arg name="paused" default="false"/>
   <arg name="use_sim_time" default="true"/>
   <arg name="gui" default="true"/>
   <arg name="headless" default="false"/>
