+++ src/gazebo_plugins_auvsim/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(gazebo_plugins)
+project(gazebo_plugins_auvsim)
 
 ## Compile as C++11, supported in ROS Kinetic and newer
   add_compile_options(-std=c++11)
